# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Portal REST API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API for Job Portal featuring role-based access control, AI-powered matching, and candidate pipelines."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["full_name", "email", "password"],
                "properties": {
                  "full_name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string", "enum": ["job_seeker", "employer"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registration successful" },
          "400": { "description": "Validation error" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate user and issue JWT",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get authenticated user profile",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile data" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "List active jobs with filters and pagination",
        "tags": ["Jobs"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "job_type", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of jobs" }
        }
      },
      "post": {
        "summary": "Post a new job listing (Employers/Admins)",
        "tags": ["Jobs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "company_name", "location", "description"],
                "properties": {
                  "title": { "type": "string" },
                  "company_name": { "type": "string" },
                  "location": { "type": "string" },
                  "job_type": { "type": "string" },
                  "category": { "type": "string" },
                  "description": { "type": "string" },
                  "salary_min": { "type": "number" },
                  "salary_max": { "type": "number" },
                  "skills": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Job created" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  }
}